extract shared C++ detection core from the N-API glue: - move CLDInput/CLDOutput and DetectLanguage() out of src/cld.cc into new src/cld_core.h/.cc, with zero Napi:: dependency - src/cld.cc now only does Napi-specific arg unpacking/repacking and calls the shared function - add src/cld_core.cc to binding.gyp, no behavior change (native test suite passes unchanged)
